--- sgv2-docsapi/src/main/java/io/stargate/sgv2/docsapi/api/service/json/DeadLeaf.java
@@ -1,13 +1,0 @@
-package io.stargate.sgv2.docsapi.api.service.json;
-
-import org.immutables.value.Value;
-
-@Value.Immutable(prehash = true)
-public interface DeadLeaf {
-  String STAR = "*";
-  String ARRAY = "[*]";
-  DeadLeaf ARRAYLEAF = ImmutableDeadLeaf.builder().name(DeadLeaf.ARRAY).build();
-  DeadLeaf STARLEAF = ImmutableDeadLeaf.builder().name(DeadLeaf.STAR).build();
-
-  String getName();
-}
