+++ sgv2-docsapi/src/main/java/io/stargate/sgv2/docsapi/service/json/NoOpDeadLeafCollector.java
@@ -1,0 +1,3 @@
+package io.stargate.sgv2.docsapi.service.json;
+
+public class NoOpDeadLeafCollector implements DeadLeafCollector {}
